+++ src/main/kotlin/code_assistant/common/Icons.kt
@@ -1,0 +1,16 @@
+package code_assistant.common
+
+import com.intellij.openapi.util.IconLoader
+
+object Icons {
+    @JvmField
+    val Default = IconLoader.getIcon("/icons/app-icon-dark.svg", javaClass)
+
+    @JvmField
+    val DefaultSmall = IconLoader.getIcon("/icons/app-icon-light.svg", javaClass)
+}
+
+
+
+
+