+++ src/main/kotlin/code_assistant/settings/OpenSettingsAction.kt
@@ -1,10 +1,9 @@
-package settings
+package code_assistant.settings
 
 import com.intellij.icons.AllIcons
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.options.ShowSettingsUtil
-
 
 class OpenSettingsAction : AnAction(
     "Open Setting",
@@ -16,4 +15,3 @@ class OpenSettingsAction : AnAction(
             .showSettingsDialog(e.project, CodeAssistantSettingsConfigurable::class.java)
     }
 }
-