+++ CHANGELOG.md
@@ -6,12 +6,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.5] - 2024-03-05
+
+### Added
+
+- Support for bytedance dorado token  (#010)
+
+### Fixed
+
+- fix settings UI bugs (#011)
+- fix input "/n" bugs  (#012)
+- fix editor action prop menus show bug (#013)
+
+## [1.0.2] - 2024-03-05
+
+### Added
+
+- Support for editor action  (#002)
+
+### Fixed
+
+- editor action custom settings  (#002)
+
 ## [1.0.0] - 2024-03-04
 
 ### Added
 
-- Support for custom ollama-compatible service (#111)
+- Support for custom ollama-compatible service (#001)
 
 ### Fixed
 
-- Ollama settings  (#111)
+- Ollama settings  (#001)